requirements.txt
asgiref==3.7.2
Django==4.2.11
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==24.0
pillow==10.2.0
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
